+++ conan/tools/cmake/__init__.py
@@ -2,3 +2,4 @@ from conan.tools.cmake.toolchain import CMakeToolchain
 from conan.tools.cmake.cmake import CMake
 from conan.tools.cmake.cmakedeps.cmakedeps import CMakeDeps
 from conan.tools.cmake.file_api import CMakeFileAPI
+from conan.tools.cmake.layout import cmake_layout, clion_layout
