+++ conan/tools/microsoft/__init__.py
@@ -3,3 +3,4 @@ from conan.tools.microsoft.msbuild import MSBuild
 from conan.tools.microsoft.msbuilddeps import MSBuildDeps
 from conan.tools.microsoft.visual import msvc_runtime_flag, VCVars, is_msvc
 from conan.tools.microsoft.subsystems import subsystem_path
+from conan.tools.microsoft.layout import vs_layout
