+++ conans/test/assets/pkg_cmake.py
@@ -12,8 +12,7 @@ def pkg_cmake(name, version, requires=None, exe=False):
     conanfile = textwrap.dedent("""\
         import os
         from conans import ConanFile
-        from conan.tools.cmake import CMake
-        from conan.tools.layout import cmake_layout
+        from conan.tools.cmake import CMake, cmake_layout
 
         class Pkg(ConanFile):
             name = "{pkg_name}"
@@ -72,8 +71,7 @@ def pkg_cmake_test(require_name):
     conanfile = textwrap.dedent("""\
         import os
         from conans import ConanFile
-        from conan.tools.cmake import CMake
-        from conan.tools.layout import cmake_layout
+        from conan.tools.cmake import CMake, cmake_layout
 
         class Pkg(ConanFile):
             settings = "os", "compiler", "arch", "build_type"
@@ -108,8 +106,7 @@ def pkg_cmake_app(name, version, requires=None):
     conanfile = textwrap.dedent("""\
         import os
         from conans import ConanFile
-        from conan.tools.cmake import CMake
-        from conan.tools.layout import cmake_layout
+        from conan.tools.cmake import CMake, cmake_layout
 
         class Pkg(ConanFile):
             name = "{pkg_name}"
