+++ conans/test/functional/layout/test_build_system_layout_helpers.py
@@ -107,7 +107,7 @@ def test_error_no_msvc():
     # https://github.com/conan-io/conan/issues/9953
     conanfile = textwrap.dedent("""
         from conans import ConanFile
-        from conan.tools.layout import cmake_layout
+        from conan.tools.cmake import cmake_layout
         class Pkg(ConanFile):
             settings = "os", "compiler", "build_type", "arch"
             def layout(self):
@@ -135,7 +135,7 @@ def test_error_no_build_type():
     # https://github.com/conan-io/conan/issues/9953
     conanfile = textwrap.dedent("""
         from conans import ConanFile
-        from conan.tools.layout import cmake_layout
+        from conan.tools.cmake import cmake_layout
         class Pkg(ConanFile):
             settings = "os", "compiler", "arch"
             def layout(self):
