+++ conans/test/functional/toolchains/cmake/cmakedeps/test_cmakedeps_and_linker_flags.py
@@ -17,8 +17,7 @@ def test_shared_link_flags():
     """
     conanfile = textwrap.dedent("""
     from conans import ConanFile
-    from conan.tools.cmake import CMake
-    from conan.tools.layout import cmake_layout
+    from conan.tools.cmake import CMake, cmake_layout
 
 
     class HelloConan(ConanFile):
