+++ conans/test/functional/toolchains/cmake/cmakedeps/test_cmakedeps_components_names.py
@@ -644,8 +644,7 @@ class TestComponentsCMakeGenerators:
         conanfile = textwrap.dedent("""
             import os
             from conans import ConanFile
-            from conan.tools.cmake import CMake
-            from conan.tools.layout import cmake_layout
+            from conan.tools.cmake import CMake, cmake_layout
 
             class Conan(ConanFile):
                 name = "consumer"
