+++ conans/test/functional/toolchains/cmake/test_cmake_and_no_soname_flag.py
@@ -20,8 +20,7 @@ def test_no_soname_flag():
     client = TestClient()
     conanfile = textwrap.dedent("""
     from conans import ConanFile
-    from conan.tools.cmake import CMakeToolchain, CMake
-    from conan.tools.layout import cmake_layout
+    from conan.tools.cmake import CMakeToolchain, CMake, cmake_layout
 
     class {name}Conan(ConanFile):
         name = "{name}"
