+++ conans/test/functional/toolchains/cmake/test_cmake_toolchain_vxworks_clang.py
@@ -30,8 +30,8 @@ def client():
     conanfile = textwrap.dedent("""
         import os
         from conans import ConanFile
-        from conan.tools.cmake import CMake
-        from conan.tools.layout import cmake_layout
+        from conan.tools.cmake import CMake, cmake_layout
+
         class Pkg(ConanFile):
             settings = "os", "compiler", "build_type", "arch"
             exports_sources = "*"
